Moved directories for question modal events into single files
