+++ js/modals/question-modal/events/hints.js
@@ -24,4 +24,13 @@ function hintButtonClick($button) {
    }
 }
 
-module.exports = hintButtonClick;
+function saveHint(hintText) {
+   'use strict';
+
+   $('.hint.active').data('hint-text', hintText);
+}
+
+module.exports = {
+   buttonClick: hintButtonClick,
+   saveHint: saveHint
+};
