--- js/modals/question-modal/events/hints/save-hint.js
@@ -1,8 +1,0 @@
-
-function saveHint(hintText) {
-   'use strict';
-
-   $('.hint.active').data('hint-text', hintText);
-}
-
-module.exports = saveHint;
