--- js/modals/question-modal/events/input-output/save-input.js
@@ -1,8 +1,0 @@
-
-function saveInput(inputText) {
-   'use strict';
-
-   $('.js-input-output-active').data('input', inputText);
-}
-
-module.exports = saveInput;
