--- js/modals/question-modal/events/multiple-choice/remove-option.js
@@ -1,11 +1,0 @@
-function removeMultipleChoiceOption() {
-   'use strict';
-
-   var $multipleChoiceOptions = $('.modal .js-mc-text-option');
-
-   if ($multipleChoiceOptions.length > 2) {
-      $multipleChoiceOptions.last().remove();
-   }
-}
-
-module.exports = removeMultipleChoiceOption;
